requirements.txt
Flask>=2.2.3
Werkzeug==2.2.2
python-dotenv~=1.2.1